+++ pages/_app.tsx
@@ -1,0 +1,6 @@
+import '../styles/global.css'
+import { AppProps } from 'next/app'
+
+export default function App({ Component, pageProps }: AppProps) {
+    return <Component {...pageProps} />
+}
