+++ pages/index.tsx
@@ -1,0 +1,59 @@
+import Head from 'next/head'
+import Layout, { siteTitle } from '../components/layout'
+import utilStyles from '../styles/utils.module.css'
+import { getSortedPostsData } from '../lib/posts'
+import LInk from 'next/link'
+import Date from '../components/date'
+import { GetStaticProps } from 'next'
+import { title } from 'process'
+import Link from 'next/link'
+
+export default function Home({
+  allPostsData
+}: {
+    allPostsData: {
+      date: string 
+      title: string 
+      id: string
+  }[]
+  }) {
+  return (
+    <Layout home>
+      <Head>
+        <title>{siteTitle}</title>
+      </Head>
+      <section className={utilStyles.headingMd}>
+        <p>Hello, I'm Milos, I'm software engineer and translater for English/Serbian.</p>
+        <p>
+          (This is a sample website - you'll be building a site like this in {' '}
+          <a href="https://nextjs.org/learn">our Next.js tutorial</a>.)
+        </p>
+      </section>
+      <section className={`${utilStyles.headingMd} ${utilStyles.padding1px}`}>
+        <h2 className={utilStyles.headingLg}>Blog</h2>
+        <ul className={utilStyles.list}>
+          {allPostsData.map(({ id, date, title }) => (
+          <li className={utilStyles.listItem} key={id}>
+            <Link href={`/posts/${id}`}>
+              <a>{title}</a>
+            </Link>
+            <br />
+            <small className={utilStyles.lightText}>
+              <Date dateString={date} />
+            </small>
+          </li>
+          ))}
+        </ul>
+      </section>
+    </Layout>
+  )
+}
+
+export const getStaticProps: GetStaticProps = async () => {
+  const allPostsData = getSortedPostsData()
+  return {
+    props: {
+      allPostsData 
+    }
+  }
+}
